yt_concate/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==4.4.2
idna==3.3
imageio==2.20.0
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.1
Pillow==9.2.0
proglog==0.1.10
python-dotenv==0.20.0
pytube==12.1.0
requests==2.28.1
tqdm==4.64.0
urllib3==1.26.11
